requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
twilio==9.2.3
python-multipart>=0.0.7
pydantic==2.9.2
pydantic-settings==2.6.1
